flatpickr(달력피커) 도입 및 로컬파일업로드 디버깅

+++ src/main/java/org/godseop/apple/repository/mapper/MemberMapper.java
@@ -10,16 +10,9 @@ import java.util.List;
 @Repository("memberMapper")
 public interface MemberMapper {
 
-    // MyBatis Return
-    // SELECT - select에 해당하는 결과
-    // INSERT - 1 (여러개일 경우도 1)
-    // UPDATE - update된 행의 개수 (없다면 0)
-    // DELETE - delete된 행의 개수 (없다면 0)
-
-
     List<Member> selectMemberListAll();
 
-    Long insertMember(Member member);
+    int insertMember(Member member);
 
     int mergeMember(Member member);
